requirements.txt
Flask
Flask-RESTful
pycryptodome

